set completed tasks

+++ index.php
@@ -54,7 +54,7 @@ $closed_tasks = $conn->query("SELECT * FROM tasks Where task_compeleted= 1");
                     <?= $row['task_name'] ?>
                     <div>
                         <a href="completeTask.php?id=<?= $row['task_id']; ?>" class="btn btn-outline-success">Complete</a>
-                        <a href="deleteTask.php?id=<?= $row['task_id']; ?>" class="btn btn-outline-danger">delete</a>
+                        <a href="deleteTask.php?id=<?= $row['task_id']; ?>" class="btn btn-outline-danger" onclick="return confirm('Do you want to delte this task ?');">delete</a>
                     </div>
                     <?php endwhile; ?>
                 <?php else : ?>
@@ -71,7 +71,7 @@ $closed_tasks = $conn->query("SELECT * FROM tasks Where task_compeleted= 1");
                 <li class="list-group-item d-flex justify-content-between align-items-center">
                     <?= $row['task_name'] ?>
                     <div>
-                        <a href="deleteTask.php?id=<?= $row['task_id']; ?>" class="btn btn-outline-warning">Delete</a>
+                        <a href="deleteTask.php?id=<?= $row['task_id']; ?>" class="btn btn-outline-warning" >Delete</a>
                     </div>
                 </li>
                 <?php endwhile ?>
